Travis: do not mess about with different target scripts to run.
We extend the default yml from buildout.plonetest,
and our script and after_success parts are executed next to those parts from the default yml.
So we are doing things double.
Instead, restrict it to just running the standard bin/test.
If someone else wants to restructure this, go ahead.